requirements.txt
asgiref==3.2.3
Django==3.0.2
django-rest-framework==0.1.0
django-simple-history==2.8.0
djangorestframework==3.11.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
